Update pathways object
`pathways` is an internal object used to generate beyodcell's `genesets`.
It has been updated with the latest signatures from MSigDB.
Also, a directory called `data-raw`, that contains the code
and data used to generate this object in order to facilitate reproducibility
has been added to the package. The source of each data file is documented in a `README`
within their corresponding folders.